#14 - Refatora o TaskService para usar DTOs para criação, recuperação, e atualizações

+++ backend/src/main/java/com/lsc/todo_app/domain/service/TaskService.java
@@ -1,32 +1,66 @@
 package com.lsc.todo_app.domain.service;
 
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import com.lsc.todo_app.api.dto.task.CreateTaskRequest;
+import com.lsc.todo_app.api.dto.task.TaskDTO;
+import com.lsc.todo_app.api.dto.task.TaskPageDTO;
+import com.lsc.todo_app.api.dto.task.TaskSumarryDTO;
+import com.lsc.todo_app.api.dto.task.UpdateTaskRequest;
+import com.lsc.todo_app.api.dto.user.UserSumarryDTO;
 import com.lsc.todo_app.domain.entity.Task;
+import com.lsc.todo_app.domain.entity.User;
+import com.lsc.todo_app.domain.entity.enums.Status;
 import com.lsc.todo_app.domain.repository.TaskRepository;
+import com.lsc.todo_app.domain.repository.UserRepository;
 
 @Service
 public class TaskService {
     
     @Autowired
     private final TaskRepository taskRepository;
+    private final UserRepository userRepository;
 
-    public TaskService(TaskRepository taskRepository) {
+    public TaskService(TaskRepository taskRepository, UserRepository userRepository) {
         this.taskRepository = taskRepository;
+        this.userRepository = userRepository;
     }
 
-    public Task createTask(Task task) {
-        return taskRepository.save(task);
+    public TaskDTO createTask(CreateTaskRequest request) {
+        User user = userRepository.findById(request.getUserId()).orElseThrow(() -> new RuntimeException("User not found with id " + request.getUserId()));
+
+        Task task = new Task();
+        task.setTitle(request.getTitle());
+        task.setDescription(request.getDescription());
+        task.setStatus(Status.PENDING);
+        task.setUser(user);
+        task.setCreatedAt(LocalDateTime.now());
+        task.setUpdatedAt(LocalDateTime.now());
+
+        taskRepository.save(task);
+
+        return new TaskDTO(task, new UserSumarryDTO(user));
     }
 
-    public Page<Task> readTasks(Pageable pageable) {
-        return taskRepository.findAll(pageable);
+    public TaskPageDTO readTasks(Pageable pageable) {
+        Page<Task> taskPage = taskRepository.findAll(pageable);
+
+        List<TaskSumarryDTO> taskSumarryDTOs = taskPage.getContent().stream()
+        .map(task -> {
+            return new TaskSumarryDTO(task);
+        }).collect(Collectors.toList());
+
+        return new TaskPageDTO(taskSumarryDTOs, taskPage.getNumber(), taskPage.getTotalPages(), taskPage.getTotalElements(), taskPage.getSize());
     }
 
-    public Task updateTask(Long id, Task requestTask) {
+    public TaskDTO updateTask(Long id, UpdateTaskRequest requestTask) {
         return taskRepository.findById(id).map(task -> {
             if (requestTask.getTitle() != null){
                 task.setTitle(requestTask.getTitle());
@@ -37,14 +71,18 @@ public class TaskService {
             if (requestTask.getStatus() != null){
                 task.setStatus(requestTask.getStatus());
             }
-            if (requestTask.getUpdatedAt() != null){
-                task.setUpdatedAt(requestTask.getUpdatedAt());
-            }
-            return taskRepository.save(task);
+            task.setUpdatedAt(LocalDateTime.now());
+
+            Task taskUpdated = taskRepository.save(task);
+
+            return new TaskDTO(taskUpdated, new UserSumarryDTO(taskUpdated.getUser()));
         }).orElseThrow(() -> new RuntimeException("Task not found with id " + id));
     }
 
     public void deleteTask(Long id) {
+        if (!taskRepository.existsById(id)){
+            throw new RuntimeException("Task not found with id " + id);
+        }
         taskRepository.deleteById(id);
     }
